+++ storage/innobase/row/row0mysql.cc
@@ -332,9 +332,8 @@ row_mysql_store_geometry(
 		String  wkt;
 
 		/** Show the meaning of geometry data. */
-		Geometry* g = Geometry::construct(&buffer,
-						 (const char*)src,
-						 src_len);
+		Geometry* g = Geometry::construct(
+			&buffer, (const char*)src, (uint32) src_len);
 
 		if (g)
 		{
@@ -373,9 +372,8 @@ row_mysql_read_geometry(
 		String  wkt;
 
 		/** Show the meaning of geometry data. */
-		Geometry* g = Geometry::construct(&buffer,
-						 (const char*)data,
-						 *len);
+		Geometry* g = Geometry::construct(
+			&buffer, (const char*) data, (uint32) *len);
 
 		if (g)
 		{
